Cannot time operation

+++ Wearable_Motion_HR_TrackerSTM32/Core/Inc/peak_detection.h
@@ -19,7 +19,7 @@
 typedef struct {
 	float BEGIN;
 	float END;
-	float TIME;
+	double TIME;
 } Timing;
 extern float step_TIME;
 extern float step_BEGIN;
